requirements.txt
Django>=4.0,<4.1
djangorestframework>=3.13.0,<3.14.0
psycopg2-binary>=2.9.0,<3.0.0
Pillow>=5.3.0,<5.4.0
flake8>=3.6.0,<3.7.0